Background probe.py with twistd; use config file.
When running with twistd command-line arguments meant for twistd would have
to be ignored. Switching to a config file is cleaner.

Config file uses the same default settings as the arguments, and allows
override.

run.sh is a utility which can start, stop, and restart the probes, and tail
the log.