Add effective network size estimate.
This is in addition to the hourly instantaneous estimate: the nodes
online at the moment. It uses the same estimation technique as the
instantanious estimate, but only uses identifiers which appear in both
the past week and the week before. This assumes that if a node is online
sometime during the week that it will show up in the samples at least
once, which may not be the case, so this estimation technique has some
problems. It will hopefully provide a sense of the number of nodes which
are often online in addition to the instantanious which is likely at
least in part "always online."