Output files to output directory.
All generated files are put into the output directory. This allows
inserting an entire directory instead of specifying files individually.
Setting persistance to reboot uses a manifest putter
(DefaultManifestPutter) which is preferable for anything that doesn't
absolutely have to be inserted accross reboots. The existing manifest
putter was exhibiting nonresponsiveness in testing.